Added test for LiveTL layout being loaded
