+++ tests/selenium/firefox.py
@@ -29,7 +29,7 @@ def get_selenium(display: bool = False) -> webdriver.Firefox:
     fp = webdriver.FirefoxProfile()
     fp.DEFAULT_PREFERENCES["frozen"]["xpinstall.signatures.required"] = False
     options = webdriver.FirefoxOptions()
-    options.headless = True
+    options.headless = not display
     browser = FirefoxDriver(
         executable_path=__platform_drivers[su.platform],
         firefox_profile=fp,
