+++ tests/selenium/routines.py
@@ -1,30 +1,6 @@
-import functools
-import time
+from wrappers import retry_every_interval, web_test
 
-
-def web_test(test):
-    @functools.wraps(test)
-    def inner(self):
-        return test(self.__class__.web)
-
-    return inner
-
-
-def retry_every_interval(times_to_try=5, wait=1):
-    def wrapper(func):
-        @functools.wraps(func)
-        def inner(*args, **kwargs):
-            args = list(args)
-            for _ in range(times_to_try - 1):
-                try:
-                    return func(*args, **kwargs)
-                except AssertionError:
-                    time.sleep(wait)
-            return func(*args, **kwargs)
-
-        return inner
-
-    return wrapper
+chilled_cow = "https://www.youtube.com/watch?v=5qap5aO4i9A"
 
 
 class TestCases:
@@ -40,14 +16,31 @@ class TestCases:
     @web_test
     @retry_every_interval()
     def test_live_button_insertion(web):
-        # Test on ChilledCow
-        chilled_cow = "https://www.youtube.com/watch?v=5qap5aO4i9A"
-        if web.current_url != chilled_cow:
-            web.get(chilled_cow)
+        go_to_website(web, chilled_cow)
         ltlbuttons = web.find_elements_by_css_selector(".liveTLBotan")
         assert len(ltlbuttons) == 2, "LiveTL buttons not inserted in live"
+
+    @web_test
+    def test_livetl_panel_layout(web):
+        go_to_website(web, chilled_cow)
+        buttons = web.find_elements_by_css_selector(".liveTLBotan")
+        buttons[0].click()
+        iframes = web.find_elements_by_css_selector("iframe")
+        assert len(iframes) == 3, "Should have 3 iframes"
+        correct_ids = ("stream", "chat", "livetl-chat")
+        for frame, idd in zip(iframes, correct_ids):
+            assert frame.get_attribute("id") == idd, "incorrect iframe id"
+        web.switch_to.frame(iframes[2])
+        assert web.find_elements_by_css_selector(
+            "#updateInfo"
+        ), "livetl-chat hasn't loaded"
 
     @classmethod
     def teardown_class(cls):
         cls.web.quit()
         del cls.web
+
+
+def go_to_website(web, url):
+    if web.current_url != url:
+        web.get(url)
